+++ src/index.go
@@ -202,7 +202,7 @@ func main() {
 }
 
 // @title Cosmos Server API
-// @version 0.22.9
+// @version 0.22.10
 // @description REST API for Cosmos Cloud server management
 // @BasePath /cosmos
 // @securityDefinitions.apikey BearerAuth